debug add get object meta from PopulateObjectStatus in core_worker.cc to ProcessMessage store.cc

+++ src/ray/object_manager/plasma/store.cc
@@ -587,7 +587,7 @@ Status PlasmaStore::ProcessMessage(const std::shared_ptr<Client> &client,
       return Status::OK();
     }
     auto allocation = entry->GetAllocation();
-    unsigned long long address = (unsigned long long) entry->GetAllocation().address;
+    unsigned long address = (unsigned long) entry->GetAllocation().address;
 
     RAY_LOG(DEBUG) << "read meta infomation of object id " << object_id << " " << entry->GetAllocation().address << " " << entry->GetObjectInfo().object_id ;
     RAY_RETURN_NOT_OK(SendMetaReply(client, address , allocation.size, allocation.device_num));
